+++ Changelog.md
@@ -1,5 +1,8 @@
 VirtualSMC Changelog
 ====================
+#### v1.2.2
+- Improve manual fan control in SMCDellSensors (SMM access is enabled even if audio is played)
+
 #### v1.2.1
 - Fix version publishing for VirtualSMC and plugins